refactor : Bridegame 패키지 이동
