+++ src/main/java/bridge/controller/BridgeController.java
@@ -1,6 +1,7 @@
 package bridge.controller;
 
 import bridge.*;
+import bridge.domain.BridgeGame;
 import bridge.view.InputView;
 import bridge.view.OutputView;
